update of extension.js
tried again to solve the bug that the same toc entries will share the same expanded states across different documents. Even with unique identifiers for each toc entries, when refresh() get called, the correct expanded status of each individual toc entries still would not be updated. It seems this is related to the _onDidChangeTreeData that if the labels are the same, then the status will not be overwritten. Need extra work here for this minor issue.